dom/base/test/useractivation/test_useractivation_has_been_activated.html

<!DOCTYPE HTML>
<html>
<head>
  <title>User activation test: has been user gesture activated</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<iframe></iframe>
<iframe></iframe>
<script>

function waitForEvent(aTarget, aEvent, aCallback) {
  return new Promise((aResolve) => {
    aTarget.addEventListener(aEvent, function listener(event) {
      aCallback(event);
      aResolve();
    }, { once: true });
  });
}

let [iframe0, iframe1] = document.querySelectorAll("iframe");

function doCheck(aDocument, aName, aHasBeenUserGestureActivated) {
  is(SpecialPowers.wrap(aDocument).hasBeenUserGestureActivated,
     aHasBeenUserGestureActivated,
     `check has-been-user-activated on the ${aName}`);
  if (aHasBeenUserGestureActivated) {
    ok(SpecialPowers.wrap(aDocument).lastUserGestureTimeStamp > 0,
       `check last-user-gesture-timestamp on the ${aName}`);
  } else {
    is(SpecialPowers.wrap(aDocument).lastUserGestureTimeStamp, 0,
       `check last-user-gesture-timestamp on the ${aName}`);
  }
}

add_task(async function checkInitialStatus() {
  doCheck(document, "top-level document", false);
  doCheck(frames[0].document, "first iframe", false);
  doCheck(frames[1].document, "second iframe", false);
});

add_task(async function triggerUserActivation() {
  // Trigger user activation on the first iframe.
  SpecialPowers.wrap(frames[0].document).notifyUserGestureActivation();
  // We should also propagate to all the ancestors.
  doCheck(document, "top-level document", true);
  doCheck(frames[0].document, "first iframe", true);
  doCheck(frames[1].document, "second iframe", false);
});

add_task(async function iframeNavigation() {
  frames[0].frameElement.src = "file_empty.html";
  await waitForEvent(frames[0].frameElement, "load", () => {});
  // We should reset the flag on iframe that navigates away from current page,
  // but the flag on its ancestor isn't changed.
  doCheck(document, "top-level document", true);
  doCheck(frames[0].document, "first iframe", false);
  doCheck(frames[1].document, "second iframe", false);
});

add_task(async function triggerUserActivationOnCrossOriginFrame() {
  // Reset the activation flag.
  SpecialPowers.wrap(document).clearUserGestureActivation();
  doCheck(document, "top-level document", false);

  // load cross-origin test page on iframe.
  frames[0].frameElement.src = "https://example.com/tests/dom/base/test/useractivation/file_iframe_user_activated.html";
  await waitForEvent(window, "message", (event) => {
    if (event.data === "done") {
      doCheck(document, "top-level document", true);
      doCheck(frames[1].document, "second iframe", false);
    } else {
      ok(false, "receive unexpected message: " + event.data);
    }
  });
});

add_task(async function propagateToSameOriginConnectedSubframe() {
  // Reset the activation flag.
  SpecialPowers.wrap(document).clearUserGestureActivation();

  // load cross-origin test page on iframe.
  iframe0.src = "https://example.com/tests/dom/base/test/useractivation/file_iframe_check_user_activation.html";
  await waitForEvent(window, "message", (event) => {
    if (event.data !== "loaded") {
      ok(false, "receive unexpected message: " + event.data);
    }
  });

  // Trigger user activation on top-level document.
  SpecialPowers.wrap(document).notifyUserGestureActivation();
  doCheck(document, "top-level document", true);
  doCheck(iframe1.contentDocument, "second iframe", true);

  iframe0.contentWindow.postMessage("get", "*");
  await waitForEvent(window, "message", (event) => {
    if (typeof event.data === "object") {
      ok(!event.data.hasBeenActivated, "check has-been-user-activated on the first iframe");
      is(event.data.lastActivationTimestamp, 0, "check last-user-gesture-timestamp on the first iframe");
    } else {
      ok(false, "receive unexpected message: " + event.data);
    }
  });
});

add_task(async function endTests() {
  // Reset the activation flag in order not to interfere following test in the
  // verify mode which would run the test using same document couple times.
  SpecialPowers.wrap(document).clearUserGestureActivation();
});

</script>
</body>
